requirements.txt
# core packages
langchain==0.3.25
langchain-core==0.3.64    
langgraph==0.4.8

# Google Gemini integration (new SDK)
langchain_google_genai==2.1.5
google-ai-generativelanguage==0.6.18
google-genai>=1.20.0           # replaces “google-generativeai”

# tooling
pytest~=8.0
python-dotenv~=1.0
streamlit==1.45.1

